implemented rating for user,pharmacy and medicine

+++ PharmacyProjectBack/src/main/java/com/example/demo/repository/UserRepository/DermatologistRepository.java
@@ -4,6 +4,6 @@ import com.example.demo.model.Users.Dermatologist;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DermatologistRepository extends JpaRepository<Dermatologist, Integer> {
+public interface DermatologistRepository extends JpaRepository<Dermatologist, Long> {
     
 }
